+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased][unreleased]
+
+### Changed
+
+-   Fixed the bug in `tfrs.layers.loss.SamplingProbablityCorrection` that logits
+    should substract the log of item probability.
+
 ## [0.4.0][2021-01-20]
 
 ### Added