+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased][unreleased]
+## [0.5.1][2021-05-14]
 
 ### Changed